Rename install_tasks.py to installation_tasks.py
Same rationale as for the install.py -> installation.py renaming.
